Make `allow_cache_upload` attr an optional bool
Summary:
`allow_cache_upload` currently exists on various rules (e.g., `cxx_library()`) and it's of type `bool`.

This is problematic because we cannot distinguish whether the value was set to `False` or whether no value was assigned at the node. We need to be able to distinguish, so that we can fallback to the toolchain value if a target does not set a value explicitly.

Here, we change the type to an optional bool and create a utility function `cxx_attrs_get_allow_cache_upload()` which performs the fallback to `False`. `cxx_attrs_get_allow_cache_upload()` will be extended later in the stack to take another default value (e.g., from the toolchain).

Reviewed By: rmaz

Differential Revision: D54582113

fbshipit-source-id: 50684f9a457066054771caddbb0c7f1be73f690d