Add support for EWDK to the prelude / msvc toolchain
Summary:
Microsoft releases an Enterprise WDK (EWDK) bundle for building drivers and applications without needing a Visual Studio installation. This is very useful for CI builds, where unpacking a zip or mounting an iso file is all that is needed to spin up a build VM, instead of needing to download, install, validate, install WDK, install SDK, repair WDK, etc. Also, there is no license requirement to use the EWDK.

This updates the `vswhere.py` file to extract the necessary binaries, include dirs, and library dirs from the environment produced by calling the `SetupBuildEnv.bat` utility in the EWDK. Once this is done, the rest of a buck2 build operates normally.

To use this, download and "install" the EWDK to the filesystem.

```
> set EWDKDIR=C:\ewdk11
> buck2 build //path/to/target
```

X-link: https://github.com/facebook/buck2/pull/591

Reviewed By: JakobDegen

Differential Revision: D54803497

Pulled By: KapJI

fbshipit-source-id: 98c4a6813fdef7efb86e0ae787d601867947a6bc