Change 'root' -> 'roots'.
Summary:
In the next diff, we are going to add support for this syntax:

```
(
  "some_link_group",
  [
    ([":A", ":B"], "intersect", None),
    ...
```

this is a departure from the current link_group syntax which assumes that the first item in the `tuple()` is a single item.

This item is stored in `GroupMapping` as `root: provider_field(Label | None)`, but it turns-out that many of the places it is passed to can take a list of `Label`.

As such, this updates `root` to become `roots: provider_field(list[Label] | None)` so that we can support the new syntax in the next diff.

Reviewed By: stepancheg

Differential Revision: D54809112

fbshipit-source-id: 329147f545800b07e1e3145ffa6be9a17ce23142