Add support to specify a module to apple core data rule
Summary:
The core_data_model rule takes the current product_name and uses it for the module parameter in `momc`, this is fine for
core data models that use a global namespace (default for objc) but it doesn't work for classes that are defined inside a
Swift library.
This diff as an optional `module` property to `apple_core_data` which is used if it is not null.

Reviewed By: chatura-atapattu

Differential Revision: D55033314

fbshipit-source-id: 183c365c1bc9baf2427761663f48a379943056b1