Respect FORCE_RLIB for implicit sysroot deps
Summary:
As discussed in D55179027, it's not clear we have a good reason to be passing `-Cprefer-dynamic=yes`. Buck always does Rust->Rust dependencies using rlib, and there isn't a reason the standard library would need to diverge from that.

Let's revisit this if we ever start using dylib for Rust->Rust dependencies i.e. if FORCE_RLIB goes away or becomes configurable on the toolchain.

https://www.internalfb.com/code/fbsource/[002deff2c4c3b317fc6946be7e7dab5ce6ee17d9]/fbcode/buck2/prelude/rust/link_info.bzl?lines=92-94

Reviewed By: JakobDegen

Differential Revision: D55220508

fbshipit-source-id: c74912bae88bed5e417173e5e05e58162215a192