Define build modes for the apple ecosystem
Summary:
To begin to consolidate our use of apple platforms that combine sdks and build modes, we need to establish a source of truth.

In this diff, we define build modes for internal and external use.

In the next diffs, I'll be replacing existing uses in the code base with these.

Reviewed By: rmaz

Differential Revision: D55152125

fbshipit-source-id: e351aeaa225ed528eb682744a64c992b9362c1f9