[buck2] Create a separate package db for the empty lib
After #69, the real dynamic library is no longer needed by template haskell, but
is used by GHCi when loading a package into the repl.

During compilation we want to avoid race conditions that might occur running
without isolation when ghc tries to use an old library file that is removed
during the action or a file that is not yet fully written.

Fixes #84