Put transitive deps of rustdoc tests under their original crate name
Summary:
In rustdoc tests, the same dynamically named crate can be both a transitive dependency and a **renamed** direct dependency, as is the case with `//common/rust/multiplatform/smc:smc`'s dependency on `//common/smc/if:if-rust` here:

https://www.internalfb.com/code/fbsource/[c2fff4ca7ebb971cacac4ea8c47469e2275d0737]/fbcode/common/rust/multiplatform/smc/TARGETS?lines=6%2C9-11%2C28%2C34%2C37-38

https://www.internalfb.com/code/fbsource/[c2fff4ca7ebb971cacac4ea8c47469e2275d0737]/fbcode/common/smc/if/TARGETS?lines=8-9%2C51%2C53-54%2C94-95%2C97-98

For the purpose of laying out transitive dependencies, these local renames (`named_deps`) are *never* relevant, but this was being handled incorrectly.

This diff fixes transitive dependencies of rustdoc tests to be placed according to the correct real crate name, which may not be known at analysis time.

Reviewed By: JakobDegen

Differential Revision: D55397344

fbshipit-source-id: 92d726bd43ad349d033d09e16ce7f36265a165f4