add intersect.
Summary:
RFC: https://fb.workplace.com/groups/3645748045672638/posts/3827886047458836/

This adds a new traversal type for `link_groups` known as `intersect`. Given N roots, this will match on targets that present in the graph of *all* the roots (and pass the filters).

This will be used by our production iOS apps so that they can declare how many of our dylibs are structures: shared containers of code for various binaries that are sent to apple (main app, extensions, app clips, etc.)

Currently, this is done by using a two-pass system to our builds that is bane of our existence. You have to run `arc dylibs enable <app>` which runs a dep-query, often taking minutes and writing the output to a BZL file which declared in a new cell `fbobjc/dylibs`.

If that cell is present, buried in our macros is logic which **rewrites the deps field of our targets (YUCK)** based on the BZL file.

With this addition, we are one step closer to getting rid of all of that.

Reviewed By: rmaz

Differential Revision: D54809113

fbshipit-source-id: 715075327f1f8c19acd499fe75fb99ad8d5e6cac