expose Clippy configuration rule and provider
Summary: expose a ClippyConfiguration provider to be used for the `clippy.json` subtarget of Rust library/binary/tests. This is made a unique provider (instead of using `DefaultInfo`) to protect against the use of `clippy.toml` files that may, for example, have not been merged with the toolchain defined `clippy.toml` file.

Reviewed By: JakobDegen

Differential Revision: D55386480

fbshipit-source-id: 4698d5b3b5fbeeeaca058ec69484a6bd98853191