Simplify relative path calculation
Summary:
These launcher scripts are hard.

First, with simpler scripts, it is easier to debug them.

Second, scripts are not supposed to do black magic like strip three first character because we assume they are `../`). Rules are only allowed to reference paths they are given and should not assume anything about actual paths.

Simplify it a little: use `parent=1` to obtain resources parent directory instead of doing it inside the script.

I suspect this was implemented before `parent=` parameter added.

Reviewed By: JakobDegen

Differential Revision: D55820440

fbshipit-source-id: 7f646470cd5ce9a6c62f545e4cb56b630a370f1b