[buck2] Do not create empty libs
Since #90 we generate separate package-db's for use with template haskell during compilation,
and consuming packages in dependencies from ghci.

When generating a `extra-libraries` stanza, ghc will try to load the given library in any case
even when not needed. Skipping the entry from the package conf file will avoid that.