[buck2] Use a package when linking haskell binaries statically
As explained in a comment in haskell.bzl, the way packages are used by the
haskell rules of the prelude is a bit strange. These are intentionally not used
at link time, all the information is in the HaskellLibraryInfo and linker
commands are constructed manually.

But this breaks with our usage of ghcWithPackages and static linking.

For ghcWithPackages, we introduced the concept of HaskellToolchainLibraries that
represent packages available from the nix toolchain. We have to add these
packages to the ghc command line when linking, in order for their libraries to
be passed to the linker.

However, when linking statically dependencies must be passed in topological
order to the linker.

When passing static libraries using `-optl` as is done in haskell_binary, the
libraries from dependent packages will always be passed before the static
archives and this will lead to many undefined symbol errors.

The way of solving this within the original prelude would be to introduce
haskell_prebuilt_library rules for every toolchain library. This would need to
include proper information about the location of the libraries, compiler and
linker flags and so on. Basically, all the information contained in the package
conf file would need to be introduced into a rule.

This change uses a different approach by constructing a package that passes
the static archives for dependencies in the `ld-flags` (similar to adding them
with `-optl`) but also adds the haskell dependencies which is to ensure these
are passed to the linker in the correct order by ghc.