add support for WKApplication
Summary:
This diff adds support for WKApplication style watch apps, which is the default from Xcode 14. This no longer requires stub binaries and extensions, we can build and link binaries as normal.

This change is mostly removing the watchkit stub attributes and special handling. The `bundle_type` attribute is now used to determine if a bundle is a watch app or not, and the attribute is hard coded in the `apple_watchos_bundle` rule.

Reviewed By: milend

Differential Revision: D55987853

fbshipit-source-id: 1dd58f10226e35eb2d3e3093a9549333018c078c