escript with config
Summary:
We add support for config files for escript (which is something not natively supported). We do that via "trampoline", that load the config before calling the original escript.

There are two cases to consider, the unbundled and bundled one. The unbundled case already had a trampoline, so we just add some more code there.

The bundled one didn't have a trampoline, so we had to add one. Because the config files, in the bundled cases, are inside the bundled, there are quite hard to access. We found that erl_prim_loader:get_file(..) was the only reliable fast way to access it, all the other explored ways consisted in unpacking the escript. As a result, we cannot rely on file:consult or equivalent, and have to parse the bin ourselves (meaning, relying on erl_scan). That requires to add one more method to the trampoline.

Reviewed By: TheGeorge

Differential Revision: D55644400

fbshipit-source-id: 57b05509b0646fcd5e5208269f4d82d96ebcdad5