toolchains: Allow overriding all toolchain attrs
Summary:
X-link: https://github.com/facebook/buck2/pull/618

We already allowed this for a couple languages, let's extend it to all of them.

I don't see what we get by banning this, and there are clearly use cases for allowing it. Also, an open source user just needed this for the CXX toolchain.

Reviewed By: scottcao

Differential Revision: D56048492

fbshipit-source-id: 731a8015adc0acdb6b6ef12dce9de70fe7adbe40