AGAIN 3.12 segv fix
Summary: Ok so this code breaks all kinds of rules of python c-api.  One its using c-types to modify a read-only object in the runtime.   Two the structure of PyVarObject is not set in stone and its changing Py_Set_Size is suppose to be used instead of updating the struct directly.   I think ref count is moving to a new location in 3.14. So this isn't really workable in > 3.12 since the thing we are patching is deprecated and the new api is completely different so a solution is not likely to be possible. It would be easier to monkeypatch converage at runtime if this performance hack is even needed anymore

Reviewed By: itamaro

Differential Revision: D56212630

fbshipit-source-id: 28c27f026878e5fa33d868f9123377b0a9e3d4d4