Make it possible to bust the pyc action caches when Cinder changes magic number
Summary:
context: S411091

we need a way to make the pyc compile action to NOT reuse cached pyc's when we know they are invalid (e.g. due to a new cinder version being rolled out with a change to the magic number).

this diff implements such a mechanism, as a temporary mitigation until a proper fix is implemented (that uses the checked in cinder version the copile pyc's)

Reviewed By: scottcao

Differential Revision: D56281219

fbshipit-source-id: 72a5ac0ac60b6ef2eac96451ccbd8ec6ec17d095