Add generated_app_labels attribute to erlang_tests() macro
Summary:
# Context
When the `erlang_tests()` macro is given `srcs`, it may generate an erlang_app target.

# Problem
At the moment, this app gets some random labels that ideally should be controlled by the caller of the macro

# This diff
We add an optional `generated_app_labels` that the caller can use to set the actual labels this app gets. For now, we use the existing set of labels when the attribute is not provided, for compatibility, but may remove this in the future.

Reviewed By: acw224

Differential Revision: D56303540

fbshipit-source-id: c4fc503d7ac8daadf1e2935adc5d1555d7929af0