Add function `try_parse_build_target_pattern`
Summary:
## Context
Right now, we only have `parse_build_target_pattern` which will cause `fail` if the parse has error, and also we cannot catch the issue in bzl/bxl. So this diff adds a new function `try_parse_build_target_pattern`, if parsing fails, it will return an error message.

It will be useful for the following diff that we distinguish between build target pattern and regex pattern in xcode project gen.

Reviewed By: chatura-atapattu

Differential Revision: D55272012

fbshipit-source-id: 20e8bff113cb882dd86b4913fdd80be7f519dc3b