Buck2 Mockingbird: Only Include Exported Deps Of Deps
Summary: At the moment we're giving the mockingbird generator the complete dependency graph to use to make mocks from. However, we only need to let mockingbird know about classes that the target we're generating mocks for knows about. If it's a dep or a dep this isn't the case so it can be ommitted to have a smaller project description json leading to faster gen time.

Reviewed By: mkoshtenko

Differential Revision: D56414727

fbshipit-source-id: 0c59b4e3b665699b7c37585ab245ca50acbece26