add IDE sourcedb build script
Summary:
Pyright's IDE functionality will rely on a bxl query to buck to obtain a sourcedb.

But Pyright doesn't have the limitation of pyre's single module name <> file mapping. For example, in Pyright, foo.bar could mean one thing in one file and another somewhere else. Pyre cannot do this. This introduced complicated [merge logic](https://docs.google.com/document/d/1-80EkDXWMFSFUh61D8s1WqD1Uot8tOOu0Mc_3vIbxDQ/edit) required for `bxl` queries. This merge logic is not necessary for Pyright. Instead, pyright will store a mapping of target -> list of sourcedbs.

This diff implements a bxl query that can provide owning targets and sourcedbs for a file. it is a map of file name -> {owning targets, sourcedb for owning targets}. the owning targets (omitted from previous pyre bxl queries) is necessary for pyright to obtain the correct sourcedb in the mapping described above.

It is an MVP and may be optimized in the following ways in the future:
- it could provide names of other files in the owning target in order for pyright to cache this and not require re-querying
- there may be minimal perf implications of duplicated sourcedbs. we can see how much the impact is once the p0 implementation is complate
- other potential improvements

Reviewed By: grievejia

Differential Revision: D56417333

fbshipit-source-id: c63ccfc5048945074b311089f1edfdba8a9a8ea1