propagate shared interface info via LinkableNode
Summary:
The previous implementation was relying on there being deps from the link group dylib top level targets to be able to collect `SharedInterfaceInfo` providers from. In the general case this is not how the build graph is constructed, there is usually only a dep on the single dependency pool target.

Refactor the link group logic to collect `SharedInterfaceInfo` providers from the `LinkableNode` provider, and collect the providers from all matched targets in the `filtered_labels_to_links_map`.

Reviewed By: milend

Differential Revision: D56308766

fbshipit-source-id: 14a7f591134d14317afc628f5c9d4402ad6722fa