Remove quotes from the absolute argsfile
Summary:
The absolute path argsfile was introduced in the D46334204 stack to address S343754. These are argsfiles designed to be used by Xcode:

https://fburl.com/code/evqfkdm4

As part of S413213, Xcode 15.3 does not handle quoted args and we need to remove them.

```
ASTReadError during full parse
error: Command failed with exit code 1
```

As a follow up to this diff, we should refactor the code to replace absolute with xcode argsfiles for clarity.

Reviewed By: Nero5023

Differential Revision: D56668014

fbshipit-source-id: 3698030fe62c33cea7199264c4724b2e5978d2ee