Maintain symlinks for versioned bundles
Summary:
Versioned bundles should be symlinked according to the following format:

/Versions/Current -> A
for filename in /Versions/A:
  /filename -> Versions/A/filename

This diff adds this as part of the assemble_bundle.py

For incremental bundles it's difficult to handle the case where we switch from Versioned to non-Versioned - e.g. we want to avoid writing files into /Headers if it's a symlink that should become a folder. For this reason I've disabled incremental bundles when switching from Versioned to non-Versioned (which of course should be rare).

Reviewed By: blackm00n

Differential Revision: D56575140

fbshipit-source-id: 3f9e2e6b3f160a46b14500683d21552956ff87a6