Back out "Back out "[fix pyc] add constraint for bundled runtime and use it to decide interpreter""
Summary:
Original commit changeset: a62a0ef32d52

Original Phabricator Diff: D56679527

The original diff got reverted because it decides bundled interpreter for target platform and uses it everywhere, thus breaking cross-platform build. For example, if the bundled interpreter is decided and built for aarch64 but the build actually happen on x86_64, the py compile step fails because it's using an interpreter built for aarch64.

The fix is to add a new field called host_interpreter_bundled and resolve it for exec platform (same as host_interpreter). The tricky part is constraints for bundled and py version are not passed to exec config so the version and bundled have to be resolved during target platform resolution. We are adding a bundled default-only toolchain bit for this purpose.

Also added tests for cross build.

Reviewed By: itamaro

Differential Revision: D56792443

fbshipit-source-id: e41255595ae834cea7ea388e1c28eba40b8686b9