Fix linker wrapper arg file vs rpath confusion
Summary:
The predicate to identify an argfile in a linker invocation was just an argument beginning with @. Rpaths also begin with @, which caused issues when attempting to enable distributed thin-lto.

Tighten the predicate to check if the `@` prefixed argument comes after either `-rpath` or `-Xlinker` rpath. Do not treat the argument as an args file in that case.

Reviewed By: rmaz

Differential Revision: D56711257

fbshipit-source-id: 7436a37a392faddfb3e4b57198c4b45f4e862164