Teach buck2 how to use pkg-config libraries
Summary:
Not all projects want to maintain something like Meta's tp2 system. Smaller projects that don't vendor every dep are more ok with using system toolchains and system dependencies. This allows users to install libraries from their package manager and (if that package is modern and vends .pc file(s)) we can use `pkg-config` to locate that liibrary's headers and linking arguments to pass to the linker.

Does this break some determinism guarantees of buck2? Yes. But it's provides an easy on-ramp to using buck2 instead of spending ages vending all the dependencies and getting them to build. Compromise!

Reviewed By: ndmitchell

Differential Revision: D57164213

fbshipit-source-id: 622ccfca6cda3c4a25989c68681f77f04e329bd6