buck2 | apple | Clamp to minimum toolchain version
Summary:
When building across the split_transition that underlies apple_xcframework, all platforms will end up with the same minimum deployment version for their target platform.

WAFlipperKit is currently built with a script that builds iOS-arm64 for version 12.0 (so that it can be deployed back to older iPhones), and macCatalyst builds set to version 14.0. Porting this to apple_xcframework is blocked becase the minimum version for macCatalyst is 13.1 (the oldest version of Catalyst) - it can't support back to 12.0, but we need 12.0 for iPhones.

When setting a minimum target to 12.0 in Xcode and building for catalyst, I can see that it passes `-target x86_64-apple-ios13.1-macabi`, i.e. it clamps the minimum version to 13.1.

Currently, when setting the minimum target to 12.0 in Buck2 and building for catalyst, the version is missing and a nonsensical build flag is passed:
`-target x86_64-apple-ios-macabi`. (i.e. missing '13.1')

This diff clamps versions for all platforms at the `toolchain_min_version`.

Reviewed By: milend

Differential Revision: D57328419

fbshipit-source-id: 0fa2d2360d0beb2427d93c7e4868d51015aeb8a2