Fix missing package in librar catalog
The package `async` was missing in the toolchain library catalog in the
context of
https://github.com/MercuryTechnologies/the-culture-repo/issues/147.
The reason seems to be that the package is listed last, and `ghc-pkg`
does not always close the list of packages with a trailing `---`.

The corresponding code is adapted to take a last package into
account even if it is not followed by `---`.