Add elp.bxl to erlang prelude
Summary:
Currently [ELP](https://github.com/WhatsApp/erlang-language-platform) uses a hard-coded buck2 query to retrieve the project model when loading a buck2 project.
This mostly works, but does not always resolve include paths correctly, and so reports spurious unresolved include file errors for files that compile with buck2
This script is an evolution of that hard-coded query. If used without the `--deps_includes true` flag it provides equivalent information.
If the flag is set, it includes extended include path information.
The extended information is done this way to allow a controlled rollout, with a fallback position if there is a problem. In time it may be discarded.

Reviewed By: TheGeorge

Differential Revision: D57279431

fbshipit-source-id: 5bd085e6cbd6b57ffe15cfe15d39912b63b7e3f5