Make linker invocation formation API more platform agnostic
Summary:
Prior to this diff, the modifies APIs used to create linker invocations behaved differently depending on platform. In particular, linker command lines did not include object files or static archives that are included in a filelist.

This diff lifts the platform specific configuration out of these APIs. Rather than requesting linker invocations in the same way for different platforms and expecting different results, we request differently for different platforms.

This allow us to re-use the linker invocation APIs in different ways for the same platform. In particular, it allows us to easily create Darwin linker invocations that use the a filelist or do not.

Reviewed By: rmaz

Differential Revision: D57069589

fbshipit-source-id: d1987bc72ea9f9a16f0a06fffca3e757b7737f44