Add sources field to JavaPackagingDep
Summary:
In order to be able to create a sources `.jar` for `android_aar`, we need to pass all the sources up through the graph.

Therefore, adding this field to `JavaPackagingDep`. It's trivial to set it for `android_prebuilt_aar` and `prebuilt_jar`, so we do. The field will be populated for Java and Kotlin libraries later in the stack.

Reviewed By: navidqar

Differential Revision: D57490995

fbshipit-source-id: 5a7df2b5a626d9f40f957ee38ae893a070b1efd3