ide bxl script takes absolute paths
Summary:
there's no need to query for root in the sourcedb bxl script if the bxl owners query will do it. this change makes calling the script easier as it now just takes absolute path as an input.

note: this will temporarily break the pyright buck integration as this changes goes out in fbsource version control and the corresponding language server change (next diff) goes out in the extension). this is ok as it hasn't been released yet.

Reviewed By: ljw1004

Differential Revision: D57382833

fbshipit-source-id: 68dc8615548631a8f35939aa11f345dccadf29cc